Several updates to gpio.h
The pin in the struct is now represented by a bit mask instead of a bit
position. This allows a struct to hold and operate on multiple pins at
once emulating direct port manipulation.
Logical versions of setValue and setDir were added since the non-logical
now expected the value to line up with the defined pin_mask.
Changed the toggle function to use PINx=1, which should be as fast as
possible.
Added some more comments